feat(core): 调整部门权限过滤逻辑并优化用户查询字段- 修改部门数据过滤方式，暂绕过上下文获取部门ID的逻辑- 更新用户查询接口中的ID字段类型以适配新结构 - 调整日志级别为debug模式便于排查问题 - 启用数据库调试模式提升开发环境可见性- 引入新的依赖库支持命令行工具及表格展示功能 - 清理无用导入包减少冗余代码体积

+++ rpc/internal/logic/user/get_user_by_username_logic.go
@@ -2,7 +2,6 @@ package user
 
 import (
 	"context"
-
 	"github.com/saas-mingyang/mingyang-admin-common/utils/pointy"
 
 	"github.com/saas-mingyang/mingyang-admin-core/rpc/ent/user"
@@ -33,7 +32,6 @@ func (l *GetUserByUsernameLogic) GetUserByUsername(in *core.UsernameReq) (*core.
 	if err != nil {
 		return nil, dberrorhandler.DefaultEntError(l.Logger, err, in)
 	}
-
 	return &core.UserInfo{
 		Nickname:     &result.Nickname,
 		Avatar:       &result.Avatar,
@@ -43,7 +41,7 @@ func (l *GetUserByUsernameLogic) GetUserByUsername(in *core.UsernameReq) (*core.
 		Mobile:       &result.Mobile,
 		Email:        &result.Email,
 		Status:       pointy.GetPointer(uint32(result.Status)),
-		Id:           pointy.GetPointer(uint64(result.ID)),
+		Id:           pointy.GetPointer(result.ID),
 		Username:     &result.Username,
 		HomePath:     &result.HomePath,
 		Description:  &result.Description,
